Backend/requirements.txt
flask
gunicorn
requests
python-dotenv
flask-cors
jsonify